[nrf fromtree] net: shell: Rename the common.h to be more unique
As the common.h is only meant to be used by the network
shell files, rename it to be more descriptive in order to
avoid possible conflicts with any other common.h file.

(cherry picked from commit 477a4a5d34042f49830fc8aff94ffc16a98cce10)
