+++ subsys/net/lib/shell/conn.c
@@ -8,7 +8,7 @@
 #include <zephyr/logging/log.h>
 LOG_MODULE_DECLARE(net_shell);
 
-#include "common.h"
+#include "net_shell_private.h"
 
 #if defined(CONFIG_NET_TCP)
 #include "tcp_internal.h"
