+++ subsys/net/lib/shell/mem.c
@@ -8,7 +8,7 @@
 #include <zephyr/logging/log.h>
 LOG_MODULE_DECLARE(net_shell);
 
-#include "common.h"
+#include "net_shell_private.h"
 
 struct ctx_info {
 	int pos;
