+++ subsys/net/lib/shell/ping.c
@@ -13,7 +13,7 @@ LOG_MODULE_DECLARE(net_shell);
 #include <zephyr/random/random.h>
 #include <zephyr/net/icmp.h>
 
-#include "common.h"
+#include "net_shell_private.h"
 
 #include "../ip/icmpv6.h"
 #include "../ip/icmpv4.h"
