+++ subsys/net/lib/shell/ppp.c
@@ -10,7 +10,7 @@ LOG_MODULE_DECLARE(net_shell);
 
 #include <zephyr/net/ppp.h>
 
-#include "common.h"
+#include "net_shell_private.h"
 
 #if defined(CONFIG_NET_L2_PPP)
 #include <zephyr/net/ppp.h>
