+++ subsys/net/lib/shell/resume.c
@@ -9,7 +9,7 @@
 #include <zephyr/pm/device.h>
 LOG_MODULE_DECLARE(net_shell);
 
-#include "common.h"
+#include "net_shell_private.h"
 
 static int cmd_net_resume(const struct shell *sh, size_t argc, char *argv[])
 {
