+++ subsys/net/lib/shell/tcp.c
@@ -10,7 +10,7 @@ LOG_MODULE_DECLARE(net_shell);
 
 #include <stdlib.h>
 
-#include "common.h"
+#include "net_shell_private.h"
 
 #if defined(CONFIG_NET_TCP) && defined(CONFIG_NET_NATIVE_TCP)
 static struct net_context *tcp_ctx;
